Externalize record index file
This is a bit of a compromise. It is great to have an algorithm determine automatically which header a record belongs to. However, there are cases when it is much easier to explicitly say which header a record should link to when, algorithmically, it might be hard to write something that's robust. A good example here is the "TOP" record. From the perspective of an algorithm trying to figure out where "TOP" should link to, it is a bit ambiguous since there is no header named "TOP" explicitly (but there are multiple headers with "TOP" in the heading name). By externalizing the file, we can explicitly set this.