Fix tables to render without errors
Apparently, multiline ASCII tables do NOT round-trip well through Pandoc in our process. It's possible this is an issue with flags to pandocs related to allowable table types or column widths but for now, just shortening the width of each table to fix the problem. Tables appear to need to end prior to column 70...