Update ERB processing: verify context variables and make more robust
We also add an additional default context variable `test_erb` which is contained within the source code of the site front page as a test to determine erb is working (you would override the value to true in your local config.yaml file).

doc/lib/template.rb is now using a more robust method for defining binding context. Note: due to a limitation in Ruby, we can’t handle uppercase variable names which are interpreted as constants and don’t work the same was as a local variable definition.