Remove self-links to parent headers
Self-links are links to a section header in the
section path. For example, in the following doc:

---
# A

## B

### C

We will now talk about A, B, C, and D.

---

From a cross-linking standpoint for the last
paragraph, A, B, and C are in the section path,
D is not. If C is linked to section C, it would be
considered "self-linking to parent". We have
removed that.